backend/requirements.txt
fastapi[standard]>=0.115.0
uvicorn[standard]>=0.34.0
websockets>=14.0
httpx>=0.28.0
python-dotenv>=1.1.0
pydantic-settings>=2.8.0
agent-framework>=1.0.0rc1
mcp[cli]>=1.9.0
azure-ai-projects>=1.0.0
azure-identity>=1.21.0
openai>=1.70.0
python-multipart>=0.0.20
